Moved serverless into root, added removeitem event to trigger as serverless s3sync was deleting config on every run
