--- serverless-ephemera/.npmignore
@@ -1,6 +1,0 @@
-# package directories
-node_modules
-jspm_packages
-
-# Serverless directories
-.serverless
